[#xx] v2/service: Gather converters in one file
